Rename ContainerRegistryBlobClient to ContainerRegistryContentClient
