+++ sdk/containerregistry/container-registry/samples-dev/deleteBlob.ts
@@ -7,7 +7,7 @@
  */
 
 import {
-  ContainerRegistryBlobClient,
+  ContainerRegistryContentClient,
   isDownloadOciImageManifestResult,
 } from "@azure/container-registry";
 import { DefaultAzureCredential } from "@azure/identity";
@@ -19,7 +19,7 @@ async function main() {
   const endpoint = process.env.CONTAINER_REGISTRY_ENDPOINT || "<endpoint>";
   const repository = process.env.CONTAINER_REGISTRY_REPOSITORY || "library/hello-world";
   // Create a new ContainerRegistryClient
-  const client = new ContainerRegistryBlobClient(
+  const client = new ContainerRegistryContentClient(
     endpoint,
     repository,
     new DefaultAzureCredential()
