+++ sdk/containerregistry/container-registry/samples-dev/downloadCustomManifest.ts
@@ -6,7 +6,7 @@
  * @azsdk-weight 3
  */
 
-import { ContainerRegistryBlobClient } from "@azure/container-registry";
+import { ContainerRegistryContentClient } from "@azure/container-registry";
 import { DefaultAzureCredential } from "@azure/identity";
 import * as dotenv from "dotenv";
 dotenv.config();
@@ -16,7 +16,7 @@ async function main() {
   // where "myregistryname" is the actual name of your registry
   const endpoint = process.env.CONTAINER_REGISTRY_ENDPOINT || "<endpoint>";
   const repository = process.env.CONTAINER_REGISTRY_REPOSITORY || "library/hello-world";
-  const client = new ContainerRegistryBlobClient(
+  const client = new ContainerRegistryContentClient(
     endpoint,
     repository,
     new DefaultAzureCredential()
