+++ sdk/containerregistry/container-registry/samples/v1-beta/typescript/src/downloadImage.ts
@@ -6,7 +6,7 @@
  */
 
 import {
-  ContainerRegistryBlobClient,
+  ContainerRegistryContentClient,
   isDownloadOciImageManifestResult,
 } from "@azure/container-registry";
 import { DefaultAzureCredential } from "@azure/identity";
@@ -24,7 +24,7 @@ async function main() {
   // where "myregistryname" is the actual name of your registry
   const endpoint = process.env.CONTAINER_REGISTRY_ENDPOINT || "<endpoint>";
   const repository = process.env.CONTAINER_REGISTRY_REPOSITORY || "library/hello-world";
-  const client = new ContainerRegistryBlobClient(
+  const client = new ContainerRegistryContentClient(
     endpoint,
     repository,
     new DefaultAzureCredential()
