+++ sdk/containerregistry/container-registry/src/content/models.ts
