+++ sdk/containerregistry/container-registry/src/index.ts
@@ -20,4 +20,4 @@ export {
   ListTagPropertiesOptions,
 } from "./registryArtifact";
 export * from "./models";
-export * from "./blob";
+export * from "./content";
